requirements.txt
fastapi==0.115.7
loguru==0.7.3
uvicorn==0.34.0
gunicorn==23.0.0
pydantic==2.10.6
python-dotenv==1.0.1
openai==1.54.3
Werkzeug==3.1.3
httpx==0.27.2
selenium==4.27.1
lxml==5.3.0
requests==2.32.3
httpx==0.27.2
beautifulsoup4==4.12.3
aio-pika==9.5.5